+++ packages/i18n-utils/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.2.0
+
+Add support for preserving trailing slash variation in `localizeUrl`
+
 ## 1.1.0
 
 Add support for `https://apps.wordpress.com` URLs